// src/types/typography-type.ts
export type FontWeightType = {
  light: 300;
  regular: 400;
  medium: 500;
  semibold: 600;
  bold: 700;
};

export type TypographyCategoryType =
  | 'h1'
  | 'h2'
  | 'h3'
  | 'h4'
  | 'large'
  | 'regular'
  | 'small'
  | 'xsmall';

export type FontWeightLabelType =
  | 'Thin'
  | 'ExtraLight'
  | 'Light'
  | 'Regular'
  | 'Medium'
  | 'SemiBold'
  | 'Bold'
  | 'ExtraBold'
  | 'Black';